pyproject.toml
[project]
name = "customer-ticket-support-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crewai>=1.10.1",
    "crewai-tools>=1.10.1",
    "fastapi>=0.135.1",
    "langchain-groq>=1.1.2",
    "litellm>=1.82.0",
    "pydantic>=2.11.10",
    "python-dotenv>=1.1.1",
    "streamlit>=1.55.0",
    "uvicorn>=0.41.0",
]
